Separate create text/shape Fns from artboard converter
